+++ crates/_future/ryo3-tokio-websockets/src/lib.rs
@@ -5,10 +5,13 @@ mod from;
 mod py_message;
 mod py_websocket;
 mod types;
-use crate::enums::PyWebSocketMessageKind;
-use crate::py_message::{PyMessageLike, PyPingPayload, PyPongPayload, PyWsMessage};
-use crate::py_websocket::{PyWebSocket, websocket};
 use pyo3::prelude::*;
+
+use crate::{
+    enums::PyWebSocketMessageKind,
+    py_message::{PyMessageLike, PyPingPayload, PyPongPayload, PyWsMessage},
+    py_websocket::{PyWebSocket, websocket},
+};
 mod constants;
 
 pub fn pymod_add(m: &Bound<'_, PyModule>) -> PyResult<()> {